financialBackend/requirements.txt

langchain
langchain-openai
pydantic
boto3
python-dotenv
streamlit
pandas
openpyxl
xlrd
PyPDF2